+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support async for loops for all checks working on for loops
 - FTP019 which finds `OSError` instantiations with `errno` constants
 - FTP132 which finds usage of ``re`` functions with strings within functions
+- Improve FTP131 to also consider loops and lambda statements
 
 ## 25.5.14.0
 - Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`